LayoutTests/streams/readable-stream-reader.html

<!DOCTYPE html>
<script src='../resources/testharness.js'></script>
<script src='../resources/testharnessreport.js'></script>
<script src='../resources/gc.js'></script>
<script>
test(function() {
    var rs = new ReadableStream({});
    rs.getReader();
    window.gc();
    assert_throws(new TypeError(), function() { rs.getReader(); }, 'old reader should still be locking a new one even after garbage collection');
}, 'Collecting a ReadableStreamReader should not unlock its stream');
</script>
